saving a recipe and not adding duplicate ingredients

+++ grocerylist/src/main/java/com/olivia/grocerylist/db/RecipeRepository.java
@@ -2,5 +2,5 @@ package com.olivia.grocerylist.db;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RecipeRepository extends JpaRepository<Recipe, Integer> {
+public interface RecipeRepository extends JpaRepository<Recipe, Long> {
 }
